+++ src/TranslateFragments.java
@@ -226,7 +226,7 @@ public class TranslateFragments extends EsperBaseVisitor<EPLOutput> {
 			EPLOutput val = new EPLOutput("0.0");
 			if (this.visitChild(VALUE) != null 
 				&& this.visitChild(VALUE).value != null) {
-				val = tryCastToFloat(this.visitChild(VALUE).value);
+				val = Misc.tryCastToFloat(this.visitChild(VALUE).value);
 				// Remove "value" child node so it does not also get mapped to params
 				this.childNodes.remove(VALUE);
 			}
@@ -312,25 +312,4 @@ public class TranslateFragments extends EsperBaseVisitor<EPLOutput> {
 		MANAGED_OBJECT /** dictionary<string, float> for assigning to the position field in a ManagedObject */
 	}
 
-	/**
-	 * Method used when translating measurement fragments. The MeasurementValue
-	 * "value" field needs to be of type float.
-	 *
-	 * If eplout is an integer, simply add ".0" to the end. If it is already a valid
-	 * float, output it unchanged.
-	 *
-	 * If it's not a numerical literal, we don't necessarily know what type it is,
-	 * so just leave it unmolested and hope it's right.
-	 */
-	private static EPLOutput tryCastToFloat(final EPLOutput value) {
-		try {
-			Double.parseDouble(value.formatOutput()); // throws if value cannot be parsed to a number
-			if (!value.contains(".")) {
-				return value.add(".0"); // convert integer value to a float
-			}
-		} catch (NumberFormatException nfe) {
-			// Do nothing - if there a type error in the resulting EPL then PS can deal with this.
-		}
-		return value;
-	}
 }