+++ src/TranslateFunction.java
@@ -16,6 +16,7 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
+import java.util.Map;
 
 public class TranslateFunction extends EsperBaseVisitor<EPLOutput> {
 	private Scope scope;
@@ -24,6 +25,42 @@ public class TranslateFunction extends EsperBaseVisitor<EPLOutput> {
 		"getString",	
 		"getNumber"
 	)));
+
+	/* ESPER_TO_EPL_INTEGER - A set of Esper integer types, used by esperCastToEplType.
+	 * Esper integral types - all are converted to EPL integer type. The end user should call .floor() if the
+	 * expression returns non float or decimal type. In such case EPL compiler alerts the user for manual edit.
+	 * */
+	private static final Set<String> ESPER_TO_EPL_INTEGER = Collections.unmodifiableSet(new HashSet<String>() {{
+		add("long");
+		add("Long");
+		add("Int");
+		add("Integer");
+		add("int");
+	}});
+
+	/* ESPER_TO_EPL_FLOAT - A set of Esper multi-precision types, used by esperCastToEplType.
+	 * Esper multi-precision types - EPL has only float type. It is safer to call .toFloat() on an stringified float.
+	 * Rest of the expressions are translated as they are.
+	 * */
+	private static final Set<String> ESPER_TO_EPL_FLOAT = Collections.unmodifiableSet(new HashSet<String>() {{
+		add("float");
+		add("Float");
+		add("double");
+		add("Double");
+		add("BigDecimal");
+		add("java.math.BigDecimal");
+		add("Number");
+	}});
+
+	/* ESPER_TO_EPL_STRING - A set of Esper string types, used by esperCastToEplType.
+	 * Esper string types - The set is created to avoid calling .toString() on string types.
+	 * Rest of the expressions are translated with ".toString()" as it is safer to call .toStirng() on most of the EPL types.
+	 * */
+	private static final Set<String> ESPER_TO_EPL_STRING = Collections.unmodifiableSet(new HashSet<String>() {{
+		add("string");
+		add("String");
+		add("java.lang.String");
+	}});
 
 	TranslateFunction(Scope scope) {
 		this.scope = scope;
@@ -40,10 +77,22 @@ public class TranslateFunction extends EsperBaseVisitor<EPLOutput> {
 	@Override
 	public EPLOutput visitFunctionCall(EsperParser.FunctionCallContext ctx) {
 		String functionName = ctx.memberLookup().getText();
+		if(scope.getSelect()== null && functionName.startsWith("find")) {
+			return EPLOutput.cannotTranslate(ctx, "find... calls outside of a select statement");
+		}
+
 		if (ESPER_FRAGMENT_FUNCTION_NAMES.contains(functionName)) {
 			return getFragment(ctx);
 		} else if(functionName.equals("findManagedObjectById")) {
 			return findManagedObjectById(ctx);
+		} else if ("findFirstAlarmBySourceAndStatusAndType".equals(functionName))  {
+			return findFirstAlarmBySourceAndStatusAndType(ctx);
+		} else if(functionName.equals("findFirstManagedObjectByType")) {
+			return findFirstManagedObjectByType(ctx);
+		} else if(functionName.equals("findOneManagedObjectByType")) {
+			return findOneManagedObjectByType(ctx);
+		} else if(functionName.equals("cast")) {
+			return esperCastToEplType(ctx);
 		} else if(ctx.getText().equals("current_timestamp()")){
 			return new EPLOutput().add("currentTime"); // This is a special case 
 		} else if(ESPER_TO_EPL_METHOD_TRANSLATIONS.containsKey(functionName)) {
@@ -157,11 +206,7 @@ public class TranslateFunction extends EsperBaseVisitor<EPLOutput> {
 
 	/** Translates a call to findManagedObjectById. Uses TranslateUnwindowedSelectClause.asyncCall to generate a use of the FindManagedObject event protocol. */
 	private EPLOutput findManagedObjectById(EsperParser.FunctionCallContext ctx) {
-		scope.getFile().addUsing("com.apama.cumulocity.Util");
-		scope.getFile().addUsing("com.apama.cumulocity.FindManagedObject");
-		scope.getFile().addUsing("com.apama.cumulocity.FindManagedObjectResponse");
-		scope.getFile().addUsing("com.apama.cumulocity.FindManagedObjectResponseAck");
-		scope.getFile().addChannelSubscription("FindManagedObjectResponse.SUBSCRIBE_CHANNEL");
+		addUsingAndChannelSubscription("ManagedObject");
 
 		String reqName = scope.getFile().uniqueVarName("fmo");
 		EPLOutput asyncBit = new EPLOutput();
@@ -178,6 +223,132 @@ public class TranslateFunction extends EsperBaseVisitor<EPLOutput> {
 		return new EPLOutput(reqName).add(".managedObject").setExprType(Type.getByEsperName("ManagedObjectCreated"));
 	}
 
+	/** Translates a call to findFirstAlarmBySourceAndStatusAndType. Uses TranslateUnwindowedSelectClause.asyncCall to generate a use of the FindAlarm event protocol. */
+	private EPLOutput findFirstAlarmBySourceAndStatusAndType(EsperParser.FunctionCallContext ctx) {
+		addUsingAndChannelSubscription("Alarm");
+
+		String reqName = scope.getFile().uniqueVarName("findAlarm");
+		EPLOutput asyncBit = new EPLOutput();
+		TranslateExpr translateExpr = new TranslateExpr(this.scope);
+		asyncBit.
+				addLine("integer " + reqName + "_req := Util.generateReqId();").
+				addLine("send FindAlarm(" + reqName + "_req, ").
+				add("{\"source\":").add(translateExpr.visit(ctx.arguments().expr(0))).
+				add(", \"status\":").add(translateExpr.visit(ctx.arguments().expr(1))).
+				add(", \"type\":").add(translateExpr.visit(ctx.arguments().expr(2))).
+				add("}) to FindAlarm.SEND_CHANNEL;");
+		EventExpression receive = new EventExpression("FindAlarmResponse", reqName);
+		receive.addConstraint("reqId", "=", reqName + "_req");
+		EventExpression terminate = new EventExpression("FindAlarmResponseAck");
+		terminate.addConstraint("reqId", "=", reqName + "_req");
+		scope.getSelect().asyncCall(asyncBit, receive, terminate);
+		return new EPLOutput(reqName).add(".alarm").setExprType(Type.getByEsperName("AlarmCreated"));
+	}
+
+	/** Translates a call to findFirstManagedObjectByType. Uses TranslateUnwindowedSelectClause.asyncCall to generate a use of the FindManagedObject event protocol. */
+	private EPLOutput findFirstManagedObjectByType(EsperParser.FunctionCallContext ctx) {
+		addUsingAndChannelSubscription("ManagedObject");
+
+		String reqName = scope.getFile().uniqueVarName("fmo");
+		String fmo = scope.getFile().uniqueVarName("fmo");
+		EPLOutput asyncBit = new EPLOutput();
+		asyncBit.
+			addLine("integer " + reqName + "_req := Util.generateReqId();").
+			addLine("FindManagedObject " + fmo + " := new FindManagedObject;").
+			addLine(fmo + ".reqId := " + reqName +"_req;").
+			addLine(fmo + ".params[\"type\"] := " + (new TranslateExpr(scope).visit(ctx.arguments().expr(0))).formatOutput() +";").
+			addLine("send " + fmo + " to FindManagedObject.SEND_CHANNEL;");
+		EventExpression receive = new EventExpression("FindManagedObjectResponse", reqName);
+		receive.addConstraint("reqId", "=", reqName + "_req");
+		EventExpression terminate = new EventExpression("FindManagedObjectResponseAck");
+		terminate.addConstraint("reqId", "=", reqName + "_req");
+		scope.getSelect().asyncCall(asyncBit, receive, terminate);
+		return new EPLOutput(reqName).add(".managedObject").setExprType(Type.getByEsperName("ManagedObjectCreated"));
+	}
+
+	/** Translates a call to findOneManagedObjectByType. Uses TranslateUnwindowedSelectClause.asyncCall to generate a use of the FindManagedObject event protocol. */
+	private EPLOutput findOneManagedObjectByType(EsperParser.FunctionCallContext ctx) {
+		addUsingAndChannelSubscription("ManagedObject");
+		String reqName = scope.getFile().uniqueVarName("fmo");
+		String fmo = scope.getFile().uniqueVarName("fmo");
+		EPLOutput asyncBit = new EPLOutput();
+		asyncBit.
+			addLine("integer " + reqName + "_req := Util.generateReqId();").
+			addLine("FindManagedObject " + fmo + " := new FindManagedObject;").
+			addLine(fmo + ".reqId := " + reqName +"_req;").
+			addLine(fmo + ".params[\"type\"] := " + (new TranslateExpr(scope).visit(ctx.arguments().expr(0))).formatOutput() +";").
+			addLine("send " + fmo + " to FindManagedObject.SEND_CHANNEL;");
+		EventExpression receive = new FindOneEventExpression("FindManagedObjectResponse", reqName);
+		receive.addConstraint("reqId", "=", reqName + "_req");
+		if(scope.getSelect()!=null) {
+			scope.getSelect().asyncCall(asyncBit, receive, null);
+		}
+		return new EPLOutput(reqName).add(".managedObject").setExprType(Type.getByEsperName("ManagedObjectCreated"));
+	}
+
+	/** Add required headers and subscribe to channel depending on predefined type for which find request is made*/
+	private void addUsingAndChannelSubscription(String type) {
+		scope.getFile().addUsing("com.apama.cumulocity.Util");
+		switch(type) {
+			case "Alarm":
+				scope.getFile().addUsing("com.apama.cumulocity.FindAlarm");
+				scope.getFile().addUsing("com.apama.cumulocity.FindAlarmResponse");
+				scope.getFile().addUsing("com.apama.cumulocity.FindAlarmResponseAck");
+				scope.getFile().addChannelSubscription("FindAlarmResponse.SUBSCRIBE_CHANNEL");
+				break;
+			case "ManagedObject":
+				scope.getFile().addUsing("com.apama.cumulocity.FindManagedObject");
+				scope.getFile().addUsing("com.apama.cumulocity.FindManagedObjectResponse");
+				scope.getFile().addUsing("com.apama.cumulocity.FindManagedObjectResponseAck");
+				scope.getFile().addChannelSubscription("FindManagedObjectResponse.SUBSCRIBE_CHANNEL");
+				break;
+		}
+	}
+
+	/** Convert Esper cast() function to corresponding EPL.
+	 * getNumber - returns float, and EPL won't allow direct coversion to integer, i.e. toInteger() is not available.
+	 *					EPL has floor() that returns integral part of float. Or round() that returns nearest integer to the float.
+	 * getString - returns string, and it is safe to place .to<type>() call on the return type.
+	 *
+	 * In all other cases, we simply return the first expression as it is.
+	 * Calling <exprType>.to<ExprType> on a type 'exprType' is forbidden in EPL. For e.g. float.toFloat() is not allowd.
+	 *
+	 * Conversion to EPL types Measurement, Operation expected to return same type.
+	 * */
+	private EPLOutput esperCastToEplType(EsperParser.FunctionCallContext ctx) {
+		String expression = ctx.arguments().expr(0).getText();
+		String esperType  = ctx.arguments().expr(1).getText();
+
+		boolean isGetNumber = expression.startsWith("getNumber");
+		boolean isGetString = expression.startsWith("getString");
+
+		// Get known expression type of Cast() first expression.
+		EPLOutput res = new EPLOutput().add(new TranslateExpr(this.scope).visit(ctx.arguments().expr(0)));
+		if (ESPER_TO_EPL_INTEGER.contains(esperType)) {
+			if (isGetNumber) {
+				return res.add(".floor()");
+			} else if (isGetString) {
+				return res.add(".toInteger()");
+			}
+		}
+
+		if (ESPER_TO_EPL_FLOAT.contains(esperType)) {
+			if (isGetString) {
+				return res.add(".toFloat()");
+			}
+		}
+
+		if (ESPER_TO_EPL_STRING.contains(esperType)) {
+			if (!isGetString) {
+				// Avoid string to string
+				return res.add(".toString()");
+			}
+		}
+
+		/* A default scenario, may need to be updateed by user. */
+		return res;
+	}
+
 	/** Some tokens will be Esper string literals - this gets out the content, minus the quoting */
 	static private String stripStringLiteral(ParserRuleContext ctx) {
 		String ret = ctx.getText();
@@ -189,3 +360,4 @@ public class TranslateFunction extends EsperBaseVisitor<EPLOutput> {
 		return null;
 	}
 }
+